수정 : 약속 추가 페이지로 약속 수정도 가능하도록 설정 ( 값 받아오기 세팅 끝)
